hr_department/requirements.txt
django==4.2
djangorestframework
psycopg2-binary
django-filter
python-dateutil
unidecode